api: move public models into api files
