+++ Sources/SpeakSwiftly/SpeakSwiftly.swift
@@ -1,5 +1,4 @@
 import Foundation
-import TextForSpeech
 
 /// The top-level namespace and startup entrypoint for the SpeakSwiftly library.
 ///
@@ -7,318 +6,6 @@ import TextForSpeech
 /// through the runtime's typed concern handles such as ``SpeakSwiftly/Runtime/generate``
 /// and ``SpeakSwiftly/Runtime/playback``.
 public enum SpeakSwiftly {
-    // MARK: Names
-
-    /// A stable operator-facing name used for stored resources such as voice profiles.
-    public typealias Name = String
-
-    /// A stable identifier for one stored text-normalization profile.
-    public typealias TextProfileID = String
-
-    /// Describes where a generation request came from and what it is related to.
-    ///
-    /// `TextForSpeech` owns the concrete model so request metadata stays identical
-    /// across normalization, generation, and downstream server surfaces.
-    public typealias RequestContext = TextForSpeech.RequestContext
-
-    /// Identifies the kind of work represented by a request.
-    public struct RequestKind: RawRepresentable, Codable, Sendable, Equatable, Hashable {
-        public static let generateSpeech = Self(rawValue: "generate_speech")
-        public static let generateAudioFile = Self(rawValue: "generate_audio_file")
-        public static let generateBatch = Self(rawValue: "generate_batch")
-        public static let getArtifact = Self(rawValue: "get_generated_file")
-        public static let listArtifacts = Self(rawValue: "list_generated_files")
-        public static let getGenerationJob = Self(rawValue: "get_generation_job")
-        public static let listGenerationJobs = Self(rawValue: "list_generation_jobs")
-        public static let expireGenerationJob = Self(rawValue: "expire_generation_job")
-        public static let createVoiceProfileFromDescription = Self(rawValue: "create_voice_profile_from_description")
-        public static let createVoiceProfileFromAudio = Self(rawValue: "create_voice_profile_from_audio")
-        public static let createSystemVoiceProfileFromDescription = Self(rawValue: "create_system_voice_profile_from_description")
-        public static let listVoiceProfiles = Self(rawValue: "list_voice_profiles")
-        public static let updateVoiceProfileName = Self(rawValue: "update_voice_profile_name")
-        public static let rerollVoiceProfile = Self(rawValue: "reroll_voice_profile")
-        public static let deleteVoiceProfile = Self(rawValue: "delete_voice_profile")
-        public static let getActiveTextProfile = Self(rawValue: "get_active_text_profile")
-        public static let getTextProfile = Self(rawValue: "get_text_profile")
-        public static let listTextProfiles = Self(rawValue: "list_text_profiles")
-        public static let getEffectiveTextProfile = Self(rawValue: "get_effective_text_profile")
-        public static let getTextProfilePersistence = Self(rawValue: "get_text_profile_persistence")
-        public static let getActiveTextProfileStyle = Self(rawValue: "get_active_text_profile_style")
-        public static let listTextProfileStyles = Self(rawValue: "list_text_profile_styles")
-        public static let setActiveTextProfileStyle = Self(rawValue: "set_active_text_profile_style")
-        public static let createTextProfile = Self(rawValue: "create_text_profile")
-        public static let updateTextProfileName = Self(rawValue: "update_text_profile_name")
-        public static let setActiveTextProfile = Self(rawValue: "set_active_text_profile")
-        public static let deleteTextProfile = Self(rawValue: "delete_text_profile")
-        public static let factoryResetTextProfiles = Self(rawValue: "factory_reset_text_profiles")
-        public static let resetTextProfile = Self(rawValue: "reset_text_profile")
-        public static let loadTextProfiles = Self(rawValue: "load_text_profiles")
-        public static let saveTextProfiles = Self(rawValue: "save_text_profiles")
-        public static let createTextReplacement = Self(rawValue: "create_text_replacement")
-        public static let replaceTextReplacement = Self(rawValue: "replace_text_replacement")
-        public static let deleteTextReplacement = Self(rawValue: "delete_text_replacement")
-        public static let listGenerationQueue = Self(rawValue: "list_generation_queue")
-        public static let listPlaybackQueue = Self(rawValue: "list_playback_queue")
-        public static let clearGenerationQueue = Self(rawValue: "clear_generation_queue")
-        public static let clearPlaybackQueue = Self(rawValue: "clear_playback_queue")
-        public static let cancelGeneration = Self(rawValue: "cancel_generation")
-        public static let cancelPlayback = Self(rawValue: "cancel_playback")
-        public static let getStatus = Self(rawValue: "get_status")
-        public static let getRuntimeOverview = Self(rawValue: "get_runtime_overview")
-        public static let getDefaultVoiceProfile = Self(rawValue: "get_default_voice_profile")
-        public static let setDefaultVoiceProfile = Self(rawValue: "set_default_voice_profile")
-        public static let getPlaybackState = Self(rawValue: "get_playback_state")
-        public static let playbackPause = Self(rawValue: "playback_pause")
-        public static let playbackResume = Self(rawValue: "playback_resume")
-        public static let setSpeechBackend = Self(rawValue: "set_speech_backend")
-        public static let reloadModels = Self(rawValue: "reload_models")
-        public static let unloadModels = Self(rawValue: "unload_models")
-        public static let clearQueue = Self(rawValue: "clear_queue")
-        public static let cancelRequest = Self(rawValue: "cancel_request")
-
-        public let rawValue: String
-
-        public init(rawValue: String) {
-            self.rawValue = rawValue
-        }
-
-        public init(from decoder: any Decoder) throws {
-            let container = try decoder.singleValueContainer()
-            rawValue = try container.decode(String.self)
-        }
-
-        public func encode(to encoder: any Encoder) throws {
-            var container = encoder.singleValueContainer()
-            try container.encode(rawValue)
-        }
-    }
-
-    /// Describes the current playback state of the live audio player.
-    public enum PlaybackState: String, Codable, Sendable, Equatable {
-        case idle
-        case playing
-        case paused
-        case interrupted
-        case recovering
-    }
-
-    /// A high-level request lifecycle event emitted by a request handle.
-    public enum RequestEvent: Sendable, Equatable {
-        case queued(QueuedEvent)
-        case acknowledged(RequestAcknowledgement)
-        case started(StartedEvent)
-        case progress(ProgressEvent)
-        case completed(RequestCompletion)
-    }
-
-    /// A point-in-time state for a submitted request.
-    public enum RequestState: Sendable, Equatable {
-        case queued(QueuedEvent)
-        case acknowledged(RequestAcknowledgement)
-        case started(StartedEvent)
-        case progress(ProgressEvent)
-        case completed(RequestCompletion)
-        case failed(Failure)
-        case cancelled(Failure)
-    }
-
-    /// Summary metrics reported while a request is being synthesized.
-    public struct SynthesisEventInfo: Sendable, Equatable {
-        public let promptTokenCount: Int
-        public let generationTokenCount: Int
-        public let prefillTime: TimeInterval
-        public let generateTime: TimeInterval
-        public let tokensPerSecond: Double
-        public let peakMemoryUsage: Double
-    }
-
-    /// A request-scoped synthesis event emitted while speech is being produced.
-    public enum SynthesisEvent: Sendable, Equatable {
-        case token(Int)
-        case info(SynthesisEventInfo)
-        case audioChunk(sampleCount: Int)
-    }
-
-    /// A meaningful event in the global generation queue.
-    public enum GenerateEvent: Codable, Sendable, Equatable {
-        case stateChanged(GenerateState)
-    }
-
-    /// Describes why global generation is not currently accepting another runnable request.
-    public enum GenerateBlockReason: String, Codable, Sendable, Equatable {
-        case waitingForResidentModel = "waiting_for_resident_model"
-        case waitingForResidentModels = "waiting_for_resident_models"
-        case waitingForActiveRequest = "waiting_for_active_request"
-        case waitingForPlaybackStability = "waiting_for_playback_stability"
-        case waitingForMarvisGenerationLane = "waiting_for_marvis_generation_lane"
-    }
-
-    /// The current semantic state of the global generation queue.
-    public enum GenerateState: Codable, Sendable, Equatable {
-        case idle
-        case running
-        case blocked(GenerateBlockReason)
-    }
-
-    /// A sequenced generation-queue state publication.
-    public struct GenerateUpdate: Codable, Sendable, Equatable {
-        public let sequence: Int
-        public let date: Date
-        public let state: GenerateState
-        public let event: GenerateEvent
-    }
-
-    /// A point-in-time read of the global generation queue.
-    public struct GenerateSnapshot: Codable, Sendable, Equatable {
-        public let sequence: Int
-        public let capturedAt: Date
-        public let state: GenerateState
-        public let activeRequests: [ActiveRequest]
-        public let queuedRequests: [QueuedRequest]
-    }
-
-    /// A meaningful event in the live playback surface.
-    public enum PlaybackEvent: Codable, Sendable, Equatable {
-        case stateChanged(PlaybackState)
-    }
-
-    /// A sequenced playback-state publication.
-    public struct PlaybackUpdate: Codable, Sendable, Equatable {
-        public let sequence: Int
-        public let date: Date
-        public let state: PlaybackState
-        public let event: PlaybackEvent
-    }
-
-    /// A point-in-time read of live playback state and queued playback work.
-    public struct PlaybackSnapshot: Codable, Sendable, Equatable {
-        public let sequence: Int
-        public let capturedAt: Date
-        public let state: PlaybackState
-        public let activeRequest: ActiveRequest?
-        public let queuedRequests: [QueuedRequest]
-        public let isRebuffering: Bool
-        public let stableBufferedAudioMS: Int?
-        public let stableBufferTargetMS: Int?
-    }
-
-    /// A meaningful event in runtime resident-model state.
-    public enum RuntimeEvent: Codable, Sendable, Equatable {
-        case stateChanged(RuntimeState)
-    }
-
-    /// The current semantic state of runtime resident-model readiness.
-    public enum RuntimeState: String, Codable, Sendable, Equatable {
-        case warmingResidentModel = "warming_resident_model"
-        case residentModelReady = "resident_model_ready"
-        case residentModelsUnloaded = "resident_models_unloaded"
-        case residentModelFailed = "resident_model_failed"
-    }
-
-    /// A sequenced runtime-state publication.
-    public struct RuntimeUpdate: Codable, Sendable, Equatable {
-        public let sequence: Int
-        public let date: Date
-        public let state: RuntimeState
-        public let event: RuntimeEvent
-    }
-
-    /// A point-in-time read of runtime resident-model and storage state.
-    public struct RuntimeSnapshot: Codable, Sendable, Equatable {
-        public let sequence: Int
-        public let capturedAt: Date
-        public let state: RuntimeState
-        public let speechBackend: SpeechBackend
-        public let residentState: ResidentModelState
-        public let defaultVoiceProfile: String
-        public let storage: RuntimeStorageSnapshot
-    }
-
-    /// A sequenced request-state update produced by the runtime's observation stream.
-    public struct RequestUpdate: Sendable, Equatable {
-        public let id: String
-        public let sequence: Int
-        public let date: Date
-        public let state: RequestState
-    }
-
-    /// A sequenced synthesis event update produced by the runtime's observation stream.
-    public struct SynthesisUpdate: Sendable, Equatable {
-        public let id: String
-        public let sequence: Int
-        public let date: Date
-        public let event: SynthesisEvent
-    }
-
-    /// A retained snapshot of the most recent known state for one request.
-    public struct RequestSnapshot: Sendable, Equatable {
-        public let id: String
-        public let kind: RequestKind
-        public let voiceProfile: String?
-        public let requestContext: RequestContext?
-        public let acceptedAt: Date
-        public let lastUpdatedAt: Date
-        public let sequence: Int
-        public let state: RequestState
-    }
-
-    /// A lightweight acknowledgement that a submitted request was accepted by the runtime.
-    public struct RequestAcknowledgement: Sendable, Equatable {
-        public let id: String
-        public let kind: RequestKind
-        public let generationJob: GenerationJob?
-    }
-
-    // MARK: Handles
-
-    /// A typed handle for one submitted request and its live event streams.
-    public struct RequestHandle: Sendable {
-        public let id: String
-        public let kind: RequestKind
-        public let voiceProfile: String?
-        public let requestContext: RequestContext?
-        /// A stream of lifecycle events such as queueing, start, progress, and completion.
-        public let events: AsyncThrowingStream<RequestEvent, any Swift.Error>
-        /// A stream of synthesis-specific updates such as token counts and audio chunks.
-        public let synthesisUpdates: AsyncThrowingStream<SynthesisUpdate, any Swift.Error>
-
-        init(
-            id: String,
-            kind: RequestKind,
-            voiceProfile: String?,
-            requestContext: RequestContext?,
-            events: AsyncThrowingStream<RequestEvent, any Swift.Error>,
-            synthesisUpdates: AsyncThrowingStream<SynthesisUpdate, any Swift.Error>,
-        ) {
-            self.id = id
-            self.kind = kind
-            self.voiceProfile = voiceProfile
-            self.requestContext = requestContext
-            self.events = events
-            self.synthesisUpdates = synthesisUpdates
-        }
-
-        /// Waits until the request reaches a terminal completion and returns its typed payload.
-        public func completion() async throws -> RequestCompletion {
-            for try await event in events {
-                if case let .completed(completion) = event {
-                    return completion
-                }
-            }
-
-            throw SpeakSwiftly.Error(
-                code: .requestCancelled,
-                message: "Request '\(id)' ended before SpeakSwiftly reported a terminal completion event.",
-            )
-        }
-    }
-
-    /// Identifies a runtime work queue for queue-specific controls.
-    public enum QueueType: String, Codable, Sendable, Equatable {
-        case generation
-        case playback
-    }
-
     // MARK: Internal Request Helpers
 
     enum SpeechJobType: Equatable {
